Fix semi activity current, to date and previous

+++ app/Rollup/Actions/SemiActivityRollup.php
@@ -167,11 +167,6 @@ class SemiActivityRollup
     {
         $actual_resources = BreakDownResourceShadow::whereIn('breakdown_resource_id', $resource_ids)->get();
 
-        $period = $this->project->open_period();
-        if (!$period) {
-            return $this->rollup_shadow;
-        }
-
         $to_date_cost = $actual_resources->sum('to_date_cost');
         $to_date_qty = $this->extra['to_date_qty'][$code] ?? 0;
         $to_date_unit_price = 0;
@@ -191,14 +186,28 @@ class SemiActivityRollup
             $this->rollup_shadow->update(compact('progress', 'status'));
         }
 
-        ActualResources::forceCreate([
+
+        $period = $this->project->open_period();
+        if (!$period) {
+            return $this->rollup_shadow;
+        }
+
+        $current_qty = $actual_resources->where('period_id', '=', $period->id)->sum('qty');
+        $current_cost = $actual_resources->where('period_id', '=', $period->id)->sum('cost');
+        $current_unit_price = $to_date_unit_price;
+        if ($current_qty) {
+            $current_unit_price = $current_cost / $current_qty;
+        }
+
+        /*ActualResources::forceCreate([
             'project_id' => $this->project->id, 'wbs_level_id' => $this->rollup_shadow->wbs_id, 'breakdown_resource_id' => $this->rollup_resource->id,
-            'qty' => $to_date_qty, 'cost' => $to_date_cost, 'unit_price' => $to_date_unit_price,
+            'qty' => $current_qty, 'cost' => $current_cost, 'unit_price' => $current_unit_price,
             'unit_id' => $this->rollup_shadow->unit_id, 'action_date' => $this->now, 'resource_id' => $this->rollup_shadow->resource_id,
             'user_id' => auth()->id(), 'batch_id' => 0, 'period_id' => $period->id, 'progress' => $progress, 'status' => $status,
         ]);
 
         ActualResources::whereIn('id', $actual_resources->pluck('id'))->where('period_id', $period->id)->delete();
+        */
 
         return $this->rollup_shadow;
     }
